Ajustes en estilos de cruds

+++ database/migrations/2023_04_29_040444_create_geographic_details_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('geographic_details', function (Blueprint $table) {
             $table->id();
-            $table->decimal('latitude');
-            $table->decimal('longitude');
+            $table->float('latitude');
+            $table->float('longitude');
             $table->string('address');
             $table->string('zip_code');
             $table->foreignId('colony_id')->constrained();
